# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateProject": {
        "properties": {
          "asset_id": {
            "pattern": "^ast_[0-9A-Za-z]+$",
            "title": "Asset Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Defaults to the asset's title.",
            "title": "Name"
          },
          "options": {
            "$ref": "#/components/schemas/ProjectOptions"
          }
        },
        "required": [
          "asset_id"
        ],
        "title": "CreateProject",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "Job": {
        "additionalProperties": false,
        "properties": {
          "attempt": {
            "default": 0,
            "minimum": 0.0,
            "title": "Attempt",
            "type": "integer"
          },
          "cache_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Key"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "pattern": "^job_[0-9A-Za-z]+$",
            "title": "Id",
            "type": "string"
          },
          "lane": {
            "enum": [
              "gpu",
              "cpu",
              "api"
            ],
            "title": "Lane",
            "type": "string"
          },
          "log_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Path"
          },
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          },
          "progress": {
            "default": 0.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "scope": {
            "additionalProperties": true,
            "default": {},
            "title": "Scope",
            "type": "object"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "default": "queued",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "canceled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "stage",
          "lane",
          "created_at"
        ],
        "title": "Job",
        "type": "object"
      },
...
      "Project": {
        "additionalProperties": false,
        "properties": {
          "asset_id": {
            "pattern": "^ast_[0-9A-Za-z]+$",
            "title": "Asset Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "pattern": "^prj_[0-9A-Za-z]+$",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "options": {
            "$ref": "#/components/schemas/ProjectOptions"
          }
        },
        "required": [
          "id",
          "asset_id",
          "name",
          "created_at"
        ],
        "title": "Project",
        "type": "object"
      },
      "ProjectDetail": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageStatus"
            },
            "title": "Stages",
            "type": "array"
          },
          "video": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video"
          }
        },
        "required": [
          "project",
          "stages",
          "video"
        ],
        "title": "ProjectDetail",
        "type": "object"
      },
      "ProjectOptions": {
        "additionalProperties": false,
        "description": "How the commentary is written and voiced. Stages read these; changing them changes the\ncreative stages' cache keys, never the analysis.",
        "properties": {
          "minutes": {
            "default": 3.0,
            "exclusiveMinimum": 0.0,
            "maximum": 240.0,
            "title": "Minutes",
            "type": "number"
          },
          "spoil_ending": {
            "default": true,
            "title": "Spoil Ending",
            "type": "boolean"
          },
          "style": {
            "default": "neutral",
            "minLength": 1,
            "title": "Style",
            "type": "string"
          },
          "voice": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice"
          }
        },
        "title": "ProjectOptions",
        "type": "object"
      },
...
      "StageStatus": {
        "properties": {
          "cached": {
            "title": "Cached",
            "type": "boolean"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "stage",
          "cached"
        ],
        "title": "StageStatus",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AI OffScreen",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs": {
      "get": {
        "description": "Newest first.",
        "operationId": "list_jobs_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "queued",
                    "running",
                    "succeeded",
                    "failed",
                    "canceled"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "lane",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "gpu",
                    "cpu",
                    "api"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lane"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "title": "Response List Jobs Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/log": {
      "get": {
        "description": "The end of the job's log file, as plain text.",
        "operationId": "get_job_log_jobs__job_id__log_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          }
        },
        "summary": "Get Job Log",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}:cancel": {
      "post": {
        "description": "Queued jobs are canceled at once; a running job stops at its next checkpoint.",
        "operationId": "cancel_job_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}:retry": {
      "post": {
        "description": "Only failed jobs can be retried.",
        "operationId": "retry_job_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          }
        },
        "summary": "Retry Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/projects": {
      "get": {
        "operationId": "list_projects_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  },
                  "title": "Response List Projects Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          }
        },
        "summary": "List Projects",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "create_project_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          }
        },
        "summary": "Create Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{project_id}": {
      "get": {
        "description": "The project, which stages are built for its options, and the finished video if any.",
        "operationId": "get_project_projects__project_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          }
        },
        "summary": "Get Project",
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{project_id}/plan:build": {
      "post": {
        "operationId": "build_plan_projects__project_id__plan_build_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          }
        },
        "summary": "Build Plan",
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{project_id}/render": {
      "post": {
        "operationId": "render_projects__project_id__render_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          }
        },
        "summary": "Render",
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/{project_id}/script:generate": {
      "post": {
        "operationId": "generate_script_projects__project_id__script_generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed in the current state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid input"
          }
        },
        "summary": "Generate Script",
        "tags": [
          "projects"
        ]
      }
    }
  }
}